Removed AI lib dependencies. Refactored the AI code in game and AI lib.
