--- AI/AI/includes/AIEvents.h
@@ -1,20 +1,0 @@
-#pragma once
-
-#include "ecsEventIncludes.h"
-
-namespace ecs
-{
-	namespace events
-	{
-		
-	}
-}
-
-//////////////////////////////////////////////
-////Event template
-////use this when creating new events
-//////////////////////////////////////////////
-//struct testEvent : public ECSEvent<testEvent>
-//{
-//	// data
-//};
