--- AI/AI/includes/AIFunctions.h
@@ -1,8 +1,0 @@
-#pragma once
-#include "ecs.h"
-
-namespace AIFunctions
-{
-	//Creates the user entities for every player and their units entities.
-	void CreatePlayerArmies(ecs::EntityComponentSystem& rEcs);
-}
