+++ AI/AI/main.cpp
@@ -1,4 +1,3 @@
-#include <ecs.h>
 
 int main() {
 	return 0;
