+++ Game/GameMain/GameMain.vcxproj
@@ -42,6 +42,7 @@
     <Import Project="..\..\PropertySheets\ModelLoader_Debug.props" />
     <Import Project="..\..\PropertySheets\InputHandler_Debug.props" />
     <Import Project="..\..\PropertySheets\Physics_Debug.props" />
+    <Import Project="..\..\PropertySheets\AI_Debug.props" />
   </ImportGroup>
   <ImportGroup Label="PropertySheets" Condition="'$(Configuration)|$(Platform)'=='Release|x64'">
     <Import Project="$(UserRootDir)\Microsoft.Cpp.$(Platform).user.props" Condition="exists('$(UserRootDir)\Microsoft.Cpp.$(Platform).user.props')" Label="LocalAppDataPlatform" />
@@ -50,6 +51,7 @@
     <Import Project="..\..\PropertySheets\ModelLoader_Release.props" />
     <Import Project="..\..\PropertySheets\Physics_Release.props" />
     <Import Project="..\..\PropertySheets\InputHandler_Release.props" />
+    <Import Project="..\..\PropertySheets\AI_Release.props" />
   </ImportGroup>
   <PropertyGroup Label="UserMacros" />
   <PropertyGroup />
@@ -86,9 +88,8 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
-    <ClCompile Include="gameAI\AIFunctions.cpp" />
-    <ClCompile Include="gameAI\GridFunctions.cpp" />
-    <ClCompile Include="gameAI\GridProp.cpp" />
+    <ClCompile Include="gameAI\AIEcsFunctions.cpp" />
+    <ClCompile Include="gameAI\GridEcsFunctions.cpp" />
     <ClCompile Include="gameUtility\CameraFunctions.cpp" />
     <ClCompile Include="Input\InterpretWebSystems.cpp" />
     <ClCompile Include="gameUtility\UtilityFunctions.cpp" />
@@ -99,11 +100,9 @@
   <ItemGroup>
     <ClInclude Include="gameAI\AIComponents.h" />
     <ClInclude Include="gameAI\AIEvents.h" />
-    <ClInclude Include="gameAI\AIFunctions.h" />
-    <ClInclude Include="gameAI\AIGlobals.h" />
+    <ClInclude Include="gameAI\AIEcsFunctions.h" />
     <ClInclude Include="gameAI\AISystems.h" />
-    <ClInclude Include="gameAI\GridFunctions.h" />
-    <ClInclude Include="gameAI\GridProp.h" />
+    <ClInclude Include="gameAI\GridEcsFunctions.h" />
     <ClInclude Include="gameAI\InitArmy.h" />
     <ClInclude Include="gameAI\InitGrid.h" />
     <ClInclude Include="gameRendering\InitMesh.h" />
