+++ Game/GameMain/gameAI/AIComponents.h
@@ -2,7 +2,7 @@
 
 #include "ecsComponentIncludes.h"
 #include <vector>
-#include "AIGlobals.h"
+#include "../../AI/AI/includes/AIGlobals.h"
 
 namespace ecs
 {
