--- Game/GameMain/gameAI/GridProp.cpp
@@ -1,3 +1,0 @@
-#include "GridProp.h"
-
-GridProp* GridProp::mpInstance = 0;
